aura-palette-backend/backend/auth_api/serializers.py
from rest_framework import serializers
from .models import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields=[
            'uid',
            'name',
            'email',
            'favPaletteIDs',
            'histPaletteIDs'
        ]